Move COCO callback to submodule.
